+++ consultation/src/main/java/com/vatunisia/soh/consultation/entity/Consultation.java
@@ -24,6 +24,6 @@ public class Consultation {
 
     private Long price;
 
-    private Long businessId;
+    private Integer businessId;
 
 }
