+++ order/src/main/java/com/vatunisia/soh/order/dto/Consultation.java
@@ -17,6 +17,6 @@ public class Consultation {
 
     private Long price;
 
-    private Long businessId;
+    private Integer businessId;
 
 }
